+++ jsonify-my-table.js
@@ -1,38 +1,32 @@
-function jsonifyMyTable(id = undefined) {
-  this.error = false;
-  const table = document.getElementById(id);
-  // 1. Does id exist in DOM?
-  if (!table) {
-    this.error = `${id} does not exist in DOM.`;
-    return;
-  }
-  // 2. Does id correspond to <table>?
-  if (table.tagName.toLowerCase() != 'table') {
-    this.error = `Element of ${id} is not a table.`;
-    return;
-  }
-  // 3. Does table have rows?
-  if ([...table.rows].length < 1) {
-    this.error = `There are no rows in table#${id}.`;
-    return;
-  }
-  // 4. Extract header.
-  this.headers = [...table.rows[0].cells].map(cell => cell.innerText);
-  // 5. Are headers unique?
-  if (this.headers.length > new Set(this.headers).size) {
-    this.error = `Headers repeat.`;
-    return;
-  }
-  // 6. Prepare the result.
-  this.res = [...table.rows].map((row, index) => {
-    if (index == 0) return; // This is the header.
-    const cells = [...row.cells].map(cell => cell.innerText);
-    return this.headers.reduce(
-      (acc, cur, index) => (acc = { ...acc, [cur.valueOf()]: cells[index] }),
-      Object.assign({})
+HTMLTableElement.prototype.jsonify = function() {
+  // 1. Does table have rows?
+  if (this.rows < 1)
+    return { error: `There are no rows in the table.` };
+
+  // 2. Extract headers.
+  const headers = [...this.rows[0].cells].map(
+    cell => cell.innerText
+  );
+
+  // 3. Are headers unique?
+  if (headers.length > new Set(headers).size)
+    return { error: `Headers repeat.` };
+
+  // 4. Prepare the result.
+  const res = [...this.rows].map((row, index) => {
+    if (index == 0) return; // This is header
+    const cells = [...row.cells].map(
+      cell => cell.innerText
+    );
+    return headers.reduce(
+      (acc, cur, index) =>
+        (acc = { ...acc, [cur.valueOf()]: cells[index] }),
+      {}
     );
   });
-  this.res.shift(); // Remove header.
-}
 
-if (!window) module.exports = jsonifyMyTable;
+  res.shift(); // Remove header
+  return { error: false, res, headers };
+};
+
+if (!window) module.exports = jsonify;