+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 2.0.0
+
+## API changes
+
+- `jsonifyMyTable` is no longer a class.
+- New function is called `jsonify`, and works on tables (`HTMLTableElement`).
+
+```javascript
+const myTable = document.getElementById('tableID');
+const data = myTable.jsonify();
+```
+
+There is no change in the result object, it contains:
+
+1. `error` if any, else set to `false`.
+2. `res` array of objects.
+3. `headers` array of headers.
+
 # 1.1.0
 
 - Eliminated `table` as the data member.